Passed main(), fixing complementary errors
Fixed the infamous xTaskGenericCreate() error by removing an extra semicolon at the end of the declaration. Proceeded forward with a few header errors and managed to arrive to 'hldlc_UT.c'. Apparently, a function defined in 'hldlc.h' called "HLDLC_frame" is defined to take 3 arguments, however the function call in 'hldlc_UT.c' uses 4 different parameters, thus returning an error. We are currently working to understand the nature of the function so we can proceed and decisively remove the extra parameter. After all, the files being compiled now are unity tests that verify the integrity of the code. Therefore, we are currently fixing the last parts of what will eventually become our first testable binary
'